+++ libcrux-ml-dsa/cg/libcrux_intrinsics_avx2.h
@@ -1,0 +1,41 @@
+/*
+ * SPDX-FileCopyrightText: 2025 Cryspen Sarl <info@cryspen.com>
+ *
+ * SPDX-License-Identifier: MIT or Apache-2.0
+ *
+ * This code was generated with the following revisions:
+ * Charon: 146b7dce58cb11ca8010b1c947c3437a959dcd88
+ * Eurydice: c06863573e1818808527b23b44e244d8b0c8e3f1
+ * Karamel: 732e3ac91245451fc441754737eef729e2b01c2a
+ * F*: 71d8221589d4d438af3706d89cb653cf53e18aab
+ * Libcrux: 26fe18b8e646819e6034de4198dc424d975b81e5
+ */
+
+#ifndef libcrux_intrinsics_avx2_H
+#define libcrux_intrinsics_avx2_H
+
+#include "eurydice_glue.h"
+#include "libcrux_mldsa_core.h"
+
+/**
+A monomorphic instance of Eurydice.arr
+with types core_core_arch_x86___m256i
+with const generics
+- $5size_t
+*/
+typedef struct Eurydice_arr_c0_s {
+  __m256i data[5U];
+} Eurydice_arr_c0;
+
+/**
+A monomorphic instance of Eurydice.arr
+with types core_core_arch_x86___m256i
+with const generics
+- $25size_t
+*/
+typedef struct Eurydice_arr_05_s {
+  __m256i data[25U];
+} Eurydice_arr_05;
+
+#define libcrux_intrinsics_avx2_H_DEFINED
+#endif /* libcrux_intrinsics_avx2_H */