+++ libcrux-ml-kem/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] (2026-01-12)
+
+- [#1280](https://github.com/cryspen/libcrux/pull/1280): Update dependencies `libcrux-platform`, `libcrux-sha3`
+- [#1276](https://github.com/cryspen/libcrux/pull/1276): (Breaking) Fix incorrect portable state serialization in incremental API (@GuuJiang)
+
 ## [0.0.4] (2025-11-05)
 
 - [#1190](https://github.com/cryspen/libcrux/pull/1190): Use secret types in kem traits